Move `fetchWrapper` into `runtime-client`
