+++ web-common/src/runtime-client/http-request-queue/HttpRequestQueue.ts
@@ -1,3 +1,7 @@
+import {
+  fetchWrapper,
+  FetchWrapperOptions,
+} from "@rilldata/web-common/runtime-client/fetchWrapper";
 import type { RequestQueueEntry } from "@rilldata/web-common/runtime-client/http-request-queue/HttpRequestQueueTypes";
 import {
   getHeapByName,
@@ -11,10 +15,6 @@ import {
   getPriority,
   InactivePriority,
 } from "@rilldata/web-common/runtime-client/http-request-queue/priorities";
-import {
-  fetchWrapper,
-  FetchWrapperOptions,
-} from "@rilldata/web-local/lib/util/fetchWrapper";
 import { appQueryStatusStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
 
 // Examples:
